requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.7
django-bootstrap4==3.0.1
django-on-heroku==1.1.2
django-registration==2.4.1
django-tinymce==3.3.0
django-utils-six==2.0
Pillow==8.3.2
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.2
whitenoise==5.3.0
